+++ src/netframework/index.ts
@@ -6,8 +6,12 @@ import iconv = require('iconv-lite');
 import moment = require('moment');
 import path = require('path');
 
+import { LoggingLevel } from './enums';
 import models = require('./models');
+import { Logger } from './services';
 import utils = require('./services/utils.service');
+
+let logger: Logger = new Logger(false, LoggingLevel.Normal);
 
 async function run() {
     try {
@@ -17,13 +21,15 @@ async function run() {
         const model = getDefaultModel();
         model.fileNames = utils.formatFileNames(model.fileNames);
 
+        logger = new Logger(model.failOnWarning, utils.mapLogLevel(model.logLevel));
+
         // Make sure path to source code directory is available
         if (!tl.exist(model.path)) {
-            tl.setResult(tl.TaskResult.Failed, `Source directory does not exist: ${model.path}`);
+            logger.error(`Source directory does not exist: ${model.path}`);
             return;
         }
 
-        utils.setCopyright(model, regExModel);
+        applyTransforms(model, regExModel);
         generateVersionNumbers(model, regExModel);
         printTaskParameters(model);
         setManifestData(model, regExModel);
@@ -32,24 +38,36 @@ async function run() {
         tl.setVariable('AssemblyInfo.FileVersion', model.fileVersion, false);
         tl.setVariable('AssemblyInfo.InformationalVersion', model.informationalVersion, false);
 
-        console.log('Complete.');
-        tl.setResult(tl.TaskResult.Succeeded, 'Complete');
+        logger.success('Complete.');
 
     } catch (err) {
-        tl.debug(err.message);
-        // tl._writeError(err);
-        // tl.setResult(tl.TaskResult.Failed, tl.loc('TaskFailed', err.message));
-        tl.setResult(tl.TaskResult.Failed, `Task failed with error: ${err.message}`);
+        logger.error(`Task failed with error: ${err.message}`);
     }
+}
+
+function applyTransforms(model: models.NetFramework, regex: models.RegEx): void {
+    Object.keys(model).forEach((key: string) => {
+        if (model.hasOwnProperty(key)) {
+            const value = Reflect.get(model, key);
+            if (typeof value === 'string' && value !== '') {
+                const newValue = utils.transformDates(value, regex);
+                if (value !== newValue) {
+                    Reflect.set(model, key, newValue);
+                    // logger.debug(`Key: ${key},  Value: ${value},  New Value: ${newValue}`);
+                }
+            }
+          }
+    });
 }
 
 function getDefaultModel(): models.NetFramework {
     const model: models.NetFramework = {
-        path: tl.getPathInput('Path', true),
+        path: tl.getPathInput('Path', true) || '',
         fileNames: tl.getDelimitedInput('FileNames', '\n', true),
         insertAttributes: tl.getBoolInput('InsertAttributes', true),
-        fileEncoding: tl.getInput('FileEncoding', true),
+        fileEncoding: tl.getInput('FileEncoding', true) || '',
         writeBOM: tl.getBoolInput('WriteBOM', true),
+
         title: tl.getInput('Title', false) || '',
         product: tl.getInput('Product', false) || '',
         description: tl.getInput('Description', false) || '',
@@ -58,11 +76,15 @@ function getDefaultModel(): models.NetFramework {
         trademark: tl.getInput('Trademark', false) || '',
         culture: tl.getInput('Culture', false) || '',
         configuration: tl.getInput('Configuration', false) || '',
+
         version: tl.getInput('VersionNumber', false) || '',
         fileVersion: tl.getInput('FileVersionNumber', false) || '',
         informationalVersion: tl.getInput('InformationalVersion', false) || '',
         verBuild: '',
         verRelease: '',
+
+        logLevel: tl.getInput('LogLevel', true) || '',
+        failOnWarning: tl.getBoolInput('FailOnWarning', true),
     };
 
     return model;
@@ -94,49 +116,54 @@ function generateVersionNumbers(model: models.NetFramework, regexModel: models.R
 
 function printTaskParameters(model: models.NetFramework): void {
 
-    console.log('Task Parameters...');
-    console.log(`\tSource folder: ${model.path}`);
-    console.log(`\tSource files: ${model.fileNames}`);
-    console.log(`\tInsert attributes: ${model.insertAttributes}`);
-    console.log(`\tFile encoding: ${model.fileEncoding}`);
-    console.log(`\tWrite unicode BOM: ${model.writeBOM}`),
+    logger.debug('Task Parameters...');
+    logger.debug(`Source folder: ${model.path}`);
+    logger.debug(`Source files: ${model.fileNames}`);
+    logger.debug(`Insert attributes: ${model.insertAttributes}`);
+    logger.debug(`File encoding: ${model.fileEncoding}`);
+    logger.debug(`Write unicode BOM: ${model.writeBOM}`),
 
-    console.log(`\tTitle: ${model.title}`);
-    console.log(`\tProduct: ${model.product}`);
-    console.log(`\tDescription: ${model.description}`);
-    console.log(`\tCompany: ${model.company}`);
-    console.log(`\tCopyright: ${model.copyright}`);
-    console.log(`\tTrademark: ${model.trademark}`);
-    console.log(`\tCulture: ${model.culture}`);
-    console.log(`\tConfiguration: ${model.configuration}`);
-    console.log(`\tAssembly version: ${model.version}`);
-    console.log(`\tAssembly file version: ${model.fileVersion}`);
-    console.log(`\tInformational version: ${model.informationalVersion}`);
-    console.log('');
+    logger.debug(`Title: ${model.title}`);
+    logger.debug(`Product: ${model.product}`);
+    logger.debug(`Description: ${model.description}`);
+    logger.debug(`Company: ${model.company}`);
+    logger.debug(`Copyright: ${model.copyright}`);
+    logger.debug(`Trademark: ${model.trademark}`);
+    logger.debug(`Culture: ${model.culture}`);
+    logger.debug(`Configuration: ${model.configuration}`);
+    logger.debug(`Assembly version: ${model.version}`);
+    logger.debug(`Assembly file version: ${model.fileVersion}`);
+    logger.debug(`Informational version: ${model.informationalVersion}`);
+
+    logger.debug(`Log Level: ${model.logLevel}`);
+    logger.debug(`Fail on Warning: ${model.failOnWarning}`);
+
+    logger.debug('');
 }
 
 function setManifestData(model: models.NetFramework, regEx: models.RegEx): void {
 
-    console.log('Setting .Net Framework assembly info...');
+    logger.info('Setting .Net Framework assembly info...');
 
     tl.findMatch(model.path, model.fileNames).forEach((file: string) => {
 
-        console.log(`Processing: ${file}`);
+        logger.info(`Processing: ${file}`);
 
         if (path.extname(file) !== '.vb' && path.extname(file) !== '.cs') {
-            console.log(`\tFile is not .vb or .cs`);
+            logger.warning(`File is not .vb or .cs`);
+            logger.info('');
             return;
         }
 
         if (!tl.exist(file)) {
-            tl.error(`File not found: ${file}`);
+            logger.error(`File not found: ${file}`);
             return;
         }
 
         setFileEncoding(file, model);
 
         if (!iconv.encodingExists(model.fileEncoding)) {
-            tl.error(`${model.fileEncoding} file encoding not supported`);
+            logger.error(`${model.fileEncoding} file encoding not supported`);
             return;
         }
 
@@ -159,8 +186,8 @@ function setManifestData(model: models.NetFramework, regEx: models.RegEx): void 
         fs.writeFileSync(file, iconv.encode(fileContent, model.fileEncoding, { addBOM: model.writeBOM, stripBOM: undefined, defaultEncoding: undefined }));
 
         const encodingResult = getFileEncoding(file);
-        console.log(`\tVerify character encoding: ${encodingResult}`);
-        console.log('');
+        logger.debug(`Verify file encoding: ${encodingResult}`);
+        logger.info('');
     });
 }
 
@@ -171,12 +198,12 @@ function getFileEncoding(file: string) {
 
 function setFileEncoding(file: string, model: models.NetFramework) {
     const encoding = getFileEncoding(file);
-    console.log(`\tDetected character encoding: ${encoding}`);
+    logger.debug(`Detected file encoding: ${encoding}`);
 
     if (model.fileEncoding === 'auto') {
         model.fileEncoding = encoding;
     } else if (model.fileEncoding !== encoding) {
-        console.log(`\tDetected character encoding is different to the one specified.`);
+        logger.warning(`Detected file encoding is different to the one specified.`);
     }
 }
 
@@ -193,7 +220,7 @@ function addUsingIfMissing(file: string, content: string) {
     usings.forEach((value, index, array) => {
         const res = content.match(new RegExp(`${value}`, 'gi'));
         if (!res || res.length <= 0) {
-            console.log(`\tAdding --> ${value}`);
+            logger.info(`Adding --> ${value}`);
             content = value.concat('\r\n', content);
         }
     });
@@ -220,7 +247,7 @@ function insertAttribute(file: string, content: string, name: string, value: str
         // ignores comments and finds correct attribute
         const res = content.match(new RegExp(`\\<Assembly:\\s*${name}`, 'gi'));
         if (!res || res.length <= 0) {
-            console.log(`\tAdding --> ${name}`);
+            logger.info(`Adding --> ${name}`);
             content += `\r\n<Assembly: ${name}("${value}")\>`;
         }
 
@@ -229,7 +256,7 @@ function insertAttribute(file: string, content: string, name: string, value: str
         // ignores comments and finds correct attribute
         const res = content.match(new RegExp(`\\[assembly:\\s*${name}`, 'gi'));
         if (!res || res.length <= 0) {
-            console.log(`\tAdding --> ${name}`);
+            logger.info(`Adding --> ${name}`);
             content += `\r\n[assembly: ${name}("${value}")\]`;
         }
     }
@@ -238,7 +265,7 @@ function insertAttribute(file: string, content: string, name: string, value: str
 }
 
 function replaceAttribute(content: string, name: string, regEx: string, value: string): string {
-    console.log(`\t${name} --> ${value}`);
+    logger.info(`${name} --> ${value}`);
     content = content.replace(new RegExp(`${name}\\s*\\w*\\(${regEx}\\)`, 'gi'), `${name}("${value}")`);
     return content;
 }