+++ ReleaseNotes.md
@@ -1,5 +1,9 @@
 # Release Notes
-> **v2.0.68**
+> **v2.1.80**
+> - Added date transforms for all fields.
+> - Enhanced logging options.
+
+> **v2.0.73**
 > - Skip invalid csproj file types.
 > - Bug fix for Net Core missing attributes.